Support type property as an array, according to JSON Schema Validation
In JSON Schema, the "type" property may be a string or an array of
strings. Previously, if type was set as array of strings (i.e., `"type":
["null", "array"]`) and there was also an "items" property defining
the array as items of type "integer," then bootprint-json-schema
would render the following:

    null,array

Additionally, the "items" property was not rendered.

The expected result for the given scenario is:

    null, integer[]

And the "items" property should be rendered with its defintion.

This supports the JSON Schema "type" as it is defined to allow an array
of strings. If one of the strings in that array is "array," then
`items.type` is used to render the datatype, and the "items" property is
also rendered.